Used priority_strategy as a starting point
