+++ src/strategies/mod.rs
@@ -1,5 +1,5 @@
 pub mod keystore;
-pub mod priority_strategy;
+pub mod uniswapx_strategy;
 pub mod shared;
 pub mod types;
-pub mod uniswapx_strategy;
+pub mod uniswapx_strategy_old;
